updated table names & attributes in schema

+++ db/migrate/20200516015416_create_journeys.rb
@@ -3,7 +3,7 @@ class CreateJourneys < ActiveRecord::Migration[6.0]
     create_table :journeys do |t|
       t.string :title
       t.string :date
-      t.integer :miles
+      t.float :miles
       t.string :location
       t.text :description
       t.integer :user_id
